feat: moved grouping logic into util function
